+++ lib/string.ex
@@ -8,6 +8,27 @@ defmodule Moar.String do
   @type string_case() :: :camel_case | :lower_camel_case | :snake_case
 
   @doc """
+  Appends `suffix` to `prefix` unless `prefix` is blank according to `Moar.Term.blank?/1`. If `suffix` is blank,
+  `prefix` is returned unchanged.
+
+  ```elixir
+  iex> Moar.String.append_unless_blank("foo", "-bar")
+  "foo-bar"
+
+  iex> Moar.String.append_unless_blank("", "-bar")
+  ""
+
+  iex> Moar.String.append_unless_blank(nil, "-bar")
+  nil
+  """
+  @spec append_unless_blank(binary() | nil, binary() | nil) :: binary()
+  def append_unless_blank(prefix, suffix) do
+    if Moar.Term.present?(prefix) && Moar.Term.present?(suffix),
+      do: prefix <> suffix,
+      else: prefix
+  end
+
+  @doc """
   Dasherizes `term`. A shortcut to `slug(term, "-")`.
 
   See docs for `slug/2`.